+++ micropython/utarfile-write/manifest.py
@@ -1,0 +1,4 @@
+metadata(description="Lightweight tarfile module writing subset", version="0.1")
+
+require("utarfile")
+package("utarfile")
